// src/modules/services/enums/service-error.messages.enum.ts
export enum ServiceErrorMessages {
  CONFLICT_NAME_VALUE_EXCEPTION = "Service with name_value already exists",
  INVALID_NAME_FORMAT = "NameValue is invalid",
  CACHE_CLEAR_ERROR = "Failed to clear cache",
  NOT_FOUND = "Service was not found",
  CREATE_ERROR = "Failed to create service",
  UPDATE_ERROR = "Failed to update service",
  DELETE_ERROR = "Failed to delete service",
}